The user can now add a member from settings
Issued in #8, the user can add members to a project(and almost remove
them, but it is not working yet).

Fixed some code comments from #19. Some are still on stand by.

As #25 asked, the settings button is just like any other menu button.

There will be probably an auto merge after this commit. Git isn't
allowing me to pull if I don't  commit.